Add 15.0, FontAwesome 4.7 and fix typos
Purpose
===========
As of version 15.0, Odoo is still using Font Awesome 4.7 for its icons.
It can hence be useful to have it as an option to have a coherent icon
with the rest of Odoo.

This commit adds support for FontAwesome 4.7, adds 15.0 to the list
(icon style has not changed since 12.0), and fixes a few typos.